Use TypeBridge to support the primitive arrays
... allowing the extra special case added in SingleRowWriter to be
backed out.

A check was earlier added cautiously in Type_coerceDatumAs, as
"safer ... than to fiddle with the canReplaceType implementations,
which could have more wide-ranging effects."

In this commit, the same check is added 30 lines later to handle
the other direction. That suggests that taking both of these checks
back out and fiddling with _Array_canReplaceType might in fact be
the right thing to do.

An oddball case is byte[]. Both Byte.c and byte_array.c define
Types that claim the Java type byte[] and map it to something
("char"[] and bytea, respectively). The bytea case is what worked
before, so the added TypeBridge ought not to break it. That leaves
the puzzle of getting it to work for "char"[]. It is such a one-off
oddball case that I am thinking a specific check for it in
coerceObjectBridged might be warranted.