Make roundtrip(,classname=>...) match getName
Change the class naming syntax accepted by TypeRoundTripper.loadClass
(and therefore by the classname parameter of roundtrip()) to be the
same as what would be output in the CLASS or CLASSJDBC columns.

It's an uglier syntax and harder to work with, but having it match
the outputs will make tests easier to write in SQL. Changing the
outputs to use the less-ugly form could be another option, but JDBC
specifies what getColumnClassName should return.

This version should also compile in Java 6. I sometimes forget
strings in switches weren't there yet.