Key s_obtainerByJavaName by name+oid for arrays
Type_fromJavaType can create a Type on the fly if it is for
an array. It is possible for one Java array type to be associated
with more than just one PostgreSQL type (for example, Java Object[]
for an array of pretty much anything, String[] and types that can
be rendered to text, and more obscure cases like byte[] being
associated with "char"[] as well as bytea).

That led to order-dependent failures when the s_obtainerByJavaName
cache was keyed only by Java name. Key it now by name+oid, but
(to limit the intrusiveness of the patch), use InvalidOid in
the key except when caching and looking up types that correspond
to Java arrays.

This C-based type hierarchy implementation has grown to what is
probably its ultimate height and wobbliness (to say nothing of how
one would even begin to audit it for faithfulness to JDBC's specified
mapping rules). My aim from here on out will be to wiggle it as little
as possible while preparing a replacement for a future major-number
update.