Make Markdown rendering work, again
The last time around this tree was 6fabe39, when the site plugin
had to be downgraded to 3.8.2 because 3.9.0 wouldn't work.

But today all the `code` markup in the Markdown documentation is
getting ignored, which seems to match the topic of this report:
https://issues.apache.org/jira/browse/DOXIA-597

which apparently can't be fixed by bumping the version of the
module that does that, so try instead eliminating the version
peg for that module and bumping the site plugin to 3.9.1. So
far, it seems that 3.9.1 doesn't fail (and still works in Maven
3.5.2) and the `code` markup works again.