Add sqlj.alias_java_language in Commands.java
It simplifies getting the details right when issuing CREATE LANGUAGE
to make an additional 'alias' for PL/Java that can be referred to
in pljava.policy for language-specific grant clauses.

It will create a sandboxed or unsandboxed language, revoke public
USAGE permission on it, and add a comment. It does not depend on
existing PL/Java LANGUAGE entries or handler functions to have their
usual names, as long as it succeeds in looking up what language(s)
and functions refer to the entry points in the loaded shared object.

Had to create an additional HTML anchor by hand in the Commands.java
javadoc for policy.md to link to, because Maven's markdown processor
corrupts URLs with fragment ids of the form generated by Javadoc
for method signatures. Did not find any documentation about getting
Maven's markdown processor not to do that.