Have examples make a java_tzset alias
Change the supplied policy to grant "user.timezone" "write" only to
functions declared in language java_tzset, rather than to everything
in the examples jar. Get the examples jar deployment to create such
a language alias and declare the issue199() function in it, so that
test succeeds, without granting the permission more broadly.

In passing, add unconditional read for a system property that is read
deep in the innards of Java 9 and 10, but that they forgot to give
themselves permission to read. Missed that in testing on more recent
Javas.