Prune pre-PG8.4 conditionals from Backend.c
The exact cutoff version for PL/Java 1.6 might not be set in stone
yet, but it will certainly not be pre-8.4.

This allows adding GUC_SUPERUSER_ONLY (which appeared in 8.4)
to some things without extra fuss. They were already only settable
by superusers, but the PostgreSQL practice seems to be to avoid
showing settings also, if they might include filesystem paths, etc.