Make validator more aggressive at class loading
In Hotspot, the validator showed the welcome behavior of rejecting
a function declaration if a dependent jar was missing; it seems
Hotspot would do enough resolution at the time of method-handle
lookup to detect such an issue.

Not so in OpenJ9, it turns out, where lazy resolution has been brought
to such a pinnacle of refinement that you can declare functions in
the S9 example class, load it, and even get their method handles, all
without the Saxon jar being present and without a peep from the validator.
(Then of course it fails when you try to call the functions, which was
what you wanted a validator to prevent in the first place.)

So tweak Function.loadClass to force initialization of the loaded class
if called with forValidator true, in an effort to front-load the
detection of about as many possible problems as can be.

As a bonus, this improves validation error reporting in Hotspot too.
Previously you could get a puzzling "unable to find method" for a
method that's obviously there, with the real problem, not finding
some class it refers to, buried in a couple layers of getCause.
Now it is reported more clearly at the time of loading the class.