Restore the trusted/untrusted language distinction
The distinction that was wiped out in 18abf79 is restored here by
introducing a special Principal, PLPrincipal, with its two concrete
subclasses Sandboxed and Unsandboxed, one of which will be in the
AccessControlContext for an executing function and can be granted
permissions in the policy file. Hence, what PostgreSQL's 'trusted'
and 'untrusted' mean for a PL can both be freely configured for
PL/Java via the policy file. The initial grants in this policy are
simply FilePermission "<<ALL FILES>>", "read,readlink,write,delete"
to Unsandboxed, and nothing special to Sandboxed. That reflects a
typical filesystem-centric view of what trusted/untrusted ought to mean,
but of course other permissions in the extensive Java permissions model
can be selectively granted as the needs of an application dictate.

The use of Sandboxed/Unsandboxed in place of PostgreSQL's trusted/
untrusted follows the usage in @Function(trust=...) since c6805e5,
discussed in [1], with the aim of being unambiguous both to people
who are or who aren't long-time users of the PostgreSQL nomenclature.

A Java Principal has both a class and a name; the PLPrincipal.Sandboxed
or .Unsandboxed in an executing function's AccessControlContext will
carry the name of the PL, as given to PostgreSQL's CREATE LANGUAGE and
used in the function's declaration. A grant in the policy file can have
the form

grant principal org.postgresql.pljava.PLPrincipal$Unsandboxed * {
  ...
};

to define permissions for any Unsandboxed function with no regard to the
name of the language, or a form like

grant principal org.postgresql.pljava.PLPrincipal$Sandboxed "java" {
  ...
};

The two CREATE LANGUAGE commands normally issued when PL/Java is
installed will create a sandboxed language named java and an unsandboxed
one named javau.

It is also possible, with additional CREATE LANGUAGE commands, to create
more named language entries that share PL/Java's handlers, and can be
used to declare Java functions. CREATE TRUSTED LANGUAGE java_net
HANDLER java_call_handler VALIDATOR java_validator; would create an
entry that could be given extra networking permissions with a grant to
PLPrincipal$Sandboxed "java_net" (and would also have any permissions
granted to PLPrincipal$Sandboxed *).

One gap in the scheme pertains to the readSQL and writeSQL methods of
"mapped UDTs", that is, user-defined types that are not implementing new
PostgreSQL base types. These methods are called by PL/Java without being
declared as functions to SQL, and therefore without SQL function
declarations to name a language or intended trust. Such methods will
execute with no PLPrincipal in their AccessControlContext, and therefore
with only the permissions the policy grants specifically to their
containing jar, or to all.

The input/output/read/write functions of a "base UDT" do have SQL
function declarations, and will be executed with the declared language
name and trust, like any ordinary function.

Recognizing this fundamental difference between the UDT types did have
the effect of simplifying a bunch of code around UDTs that had grown
a bit hairy, especially in bbbb75c.

TransactionListeners and SavepointListeners will execute with the
AccessControlContext that was in force when they were registered,
although their permissions may be limited by intersection with those
the policy grants to PL/Java itself. If such a listener needs to
exercise a privilege that is not granted to PL/Java, it will have to
wrap the exercise in doPrivileged.

[1] https://www.postgresql.org/message-id/56759A57.9060106%40anastigmatix.net