Build lids with principals
The fact that a SubjectDomainCombiner does not combine its Subject
into the domains of the /inherited/ AccessControlContext leads to
behavior that would strain the principle of least astonishment:
an SQL declaration of, say, a Java API function would execute as if
it lacked the permissions granted to the Principal that should clearly,
by inspection of the SQL declaration, be supplied.

One way to tell when you've violated the POLA is when the effort to
write a documentation paragraph that explains how the thing /does/
behave is more painful than just making it behave the way you'd
expect.

The lidWithPrincipals method could perhaps be more efficient, but
the case of SQL declarations for functions in non-PL/Java-managed
jars isn't expected to be common.