Rethought issue #312 approach
The previous approach, replacing the punctuation check with a messy
non-XML-NameChar check, persisted in an ill-thought-out belief that it's
important to avoid giving a pass on error reporting to an element name
that happens to contain the name of our wrapping element. Anybody who
wanted to sneak an element past the error reporting would not be
prevented from using exactly the name of our wrapping element to do it.
So the extra work to avoid the containing-name case achieves nothing.

Better to just count the exceptions. If we are using a wrapping element,
it will be first. No special treatment after that one (or at all, if we
are not wrapping).