Fix some lint warnings in the S9 example
Though the normal build is without -Xlint, the unchecked warning
summary message was showing up in CI build logs.

Some of the iterator games can be simplified by checking size()
of an XdmValue, generally in a context where it's known to be
zero-or-one, so not an expensive test.

Add javadoc for one method where it isn't inherited because the
source is in a different subproject.