Add control file in preparation for next release
Now that 1.6.0 is released, the next release should include
an extension SQL file allowing upgrade from 1.6.0.

Advance version to 1.6-SNAPSHOT. This is a change from past practice,
which always updated to 1.x.<next micro version>-SNAPSHOT,
which always created a change in every POM that would have to be
resolved out when merging up to master. This way will require the
same resolution tedium once, but not for future 1.6.x releases.
1.6-SNAPSHOT will just always mean a snapshot on the way to whatever
will be released next in the 1.6 line.