Add an Aggregate annotation
This can represent most of what a PostgreSQL CREATE AGGREGATE command
can say. Omitted for the moment are SERIALFUNC and DESERIALFUNC (there
isn't much PL/Java functions can do with datatype internal), and SORTOP
(not enough awareness of operators yet). There is still @SQLAction for
any rare need to get a CREATE AGGREGATE into a deployment descriptor
if this annotation can't describe it.

Since the operations of a PostgreSQL aggregate correspond nicely to
those of a Java java.util.stream.Collector (accumulator, combiner,
finisher) I've elected to use accumulate, combine, finish (plus remove)
as the annotation elements, rather than the PostgreSQLisms SFUNC/MSFUNC,
FINALFUNC, MINVFUNC, etc., ideally to spark a flash of recognition in
Java users. An SQL generator is just the thing to produce the rather
JCL-esque PostgreSQL forms from annotations that are easier to read.

For now, the only allowed target is a type, with no more significance
than that a bunch of these annotations could be conveniently hung at the
top of some source file for a class. Because an aggregate can refer to
more than one other function for its supporting operations, it did not
seem quite natural to assign a meaning to hanging the annotation on a
method. On the other hand, the only required function is the
"accumulate" of the main Plan, and the simplest aggregates can be
declared with nothing more than name, args, stateType, and accumulate;
perhaps it would be convenient some day to allow a super-shorthand where
the annotation goes on a method that could serve as accumulate and as
much as possible is inferred.

Because of the close mapping between a PostgreSQL aggregate and a Java
Collector, it is also intriguing to imagine placing this aggregate on
some class that implements Collector and trying to generate all the glue
to make a PostgreSQL aggregate of it. Intriguing, but highly
speculative, and certainly not on any roadmap at the moment.