Add an out= element to the Function annotation
Between creating a named composite type and declaring a function
to return that (it is then easy to use the function and work with
its result, but sometimes it's a nuisance to have to create a type
just for that purpose) and simply declaring RECORD as the return
type (avoids one nuisance, but then every use of the function in
SQL must be followed by AS (column definition list)), a happy medium
is to declare a function to SQL with OUT parameters. The function
can be called (without a column definition list) and produce a
result with usefully named and typed columns.

On discovering that PL/Java's runtime needs no changes to handle
functions declared that way, it seems quite useful to add annotation
support for producing that style of function declaration.