Simplify slightly and improve examples
The only difference between the function signatures from plan
to movingPlan is a possibly different stateType; no need to
re-create the whole signature when it is a mutable array. (Yes,
a clone is taken in DependTag.Function().)

Examples had a mistake in the stateSize value and did not explain
how to get it.