Generate synthetic-function declarations
It is often necessary to supply several very similar operators (and the
functions behind them), for example to complete a btree operator class.
The needed operators are simply related to each other by parameter
commutation and/or return-value negation, as already declared with
commutator and negator in the Operator annotation.

Give the Operator annotation a new element synthetic= that takes a
function name. Say there is already

@Function(name = "lt")
@Operator(name = "<")
public static void boolean less(int a, int b) { return a < b; }

Allow related operators to be declared in the same place, like this:

@Function(name = "lt")
@Operator(name = "<",  commutator = ">", negator = ">=")
@Operator(name = ">",  synthetic = "gt", negator = "<=")
@Operator(name = ">=", synthetic = "ge", commutator = "<=")
@Operator(name = "<=", synthetic = "le")
public static void boolean less(int a, int b) { return a < b; }

The order of the Operator annotations doesn't matter, only that they
include enough commutator/negator information to form a graph (rooted at
the 'real' operator) that reaches all the synthetic ones. Back edges are
added automatically (so > does not need to specify its commutator,
>= does not need to specify its negator, and <= needn't mention either).

The deployment descriptor will include CREATE FUNCTION commands for
lt, gt, ge, and le, all pointing to the same Java less() method, with
the three synthetic ones including a prefix in the AS string indicating
to commute, negate, or both. The CREATE OPERATOR commands then use the
corresponding named functions in the ordinary way.

The PL/Java runtime does not yet know what to do with the extra prefix
in a function's AS string.