Allow another shorthand notation, TWIN
Where commutator=SELF strictly means that an operator, with both
operand types the same, is its own commutator, use commutator=TWIN
on an operator whose operand types differ, to indicate that the
same-named operator with the types reversed is the commutator.

PostgreSQL itself may not even need to know that (as the names are
the same, it would just pick whichever one matched the order of the
operands in a given expression), but PL/Java's SQL generator will
use that information, if the other Operator is marked synthetic,
to know how to synthesize it.

TWIN is also a valid value for the synthetic= attribute, and means
the SQL function generated to back the operator should have the same
name as the one it is based on, which is sensible if the function
is a commutative one like multiply or add.