Add deployment-time tests
The SQLAction performing the tests must be emitted after all of the
operators it refers to. Until now, DDRProcessor has strictly enforced an
unnecessary rule that a given dependency tag must have no more than one
providing snippet. Perhaps that has helped someone detect once or twice
when code was copy/pasted without renaming a tag, but it was also quite
inconvenient for a case like this: it would require making up n
different tags for the n operators to 'provide', and the SQLAction would
then have to 'require' all n of them.

So, away with the unnecessary rule. An explicit dependency tag can be
'provided' by more than one snippet, all of which will have to be
emitted before whatever 'requires' it. The correct ordering of the test
block here can thus be declared using only one tag.

An implicit tag being 'provided' by more than one snippet will still be
reported as an error; that shouldn't happen.