Add missing bits of Aggregate annotation
This isn't the PR for aggregate generation, but that one left
a few bits unimplemented (waiting for this PR to add operator
support so sortOperator's dependency can be declared). So complete
the Aggregate annotation support for sortOperator, serialize, and
deserialize.

There is still no PL/Java support for manipulating a stateType of
internal, and hence nothing very useful to do with serialize and
deserialize, but that's no pressing reason to leave their SQL
generation unimplemented. An Aggregate annotation could be used to
create an aggregate that uses some functions in some other PL.

Add a trivial aggregate in ComplexScalar that uses sortOperator.
Check for some more errors at compile time.