Fix a copy/pasto and other minor doc fixes
The copy/pasto (magnitudeNE rather than componentsNE) had no functional
effect, but was a bit sloppy.

Some of the annotations (including some of the oldest original ones)
were missing @Documented, and therefore not showing up in javadocs.
(Aggregate and BaseUDT are used in ComplexScalar, edited in this PR,
so those changes have an excuse to be included here. MappedUDT and
Trigger aren't, but it's the same change.)

I am tempted by the idea of removing @Documented from @SQLAction ...
its values are usually longwinded, and mess up the formatting of
javadocs, and might just be better looked up in the source when
somebody wants to know. But not making that change just now.