Fix some broken doc links
Some result from modularizing code for Java 9, which adds one new
directory level in apidocs URLs, and some required adding new HTML
anchors by hand in the Javadocs where they are linked into from
Markdown docs, because Maven's Markdown processor insists on
corrupting URLs of the form Javadoc generates for method details,
as noted earlier in 5565a3c. This is irritating and hardly scalable,
but for now there aren't that many places, so the anchors are added.

It's possible Maven's Markdown processor doesn't get 100% of the
blame; it looks as if Javadoc 15 produces different fragment IDs
for method signatures than whichever previous version did that was
used when these links were made.

In passing, noticed that the noun was missing from several of the
descriptions in Commands.java: "The replace_jar will replace ..."
and so on. That raises the question what noun to use. Command?
Function? 9075-13 calls them 'procedures', which isn't entirely
satisfying (some of them return stuff), but it's the standard,
so 'procedures' it is.