Better workaround for javac 10 and 11 issue
Revert the smelly workaround added in c763cee for javac 10 and 11
misbehaving when passed a --release naming an earlier release, and
add a better workaround in DDRProcessor itself. Now it is possible
to build for --release 9 whichever toolchain is used for the build,
and that was the original intent and so is much more satisfying.

It turned out the same issue (unsurprisingly, in retrospect) would
also bite people building user code, and to have no better answer
for them than "fudge your target release if building with 10 or 11"
would be hard to stomach.

Addresses #328.