Evolve transformXML example with adjust & indent
Although XSLT 1.0 is very much like punishment when a modern version
is available with Saxon, it does have the advantage of being included
in Java rather than a large separate download. Now that it's usable,
the example functions may as well be polished a bit (say, by making
the uninteresting-but-for-testing 'how' parameters optional, and
accepting an 'adjust' parameter) and made suitable for day-to-day use.

One useful feature of the XSLT 1.0 transformer is the ability to indent
XML: great for readability, not available in core PostgreSQL, and
easy to do here. It doesn't require any particular transformation
defined; the default identity transform from the no-argument
TransformerFactory.newTransformer is enough. So make it possible to pass
null for transformName for a plain identity transform, and add optional
indent and indentWidth direct arguments to make it simple.

This can also serve as an example to clarify just how one gets xalan
to indent, as the details are subtle enough to have needed hashing out
on Stack Overflow [1].

[1] https://stackoverflow.com/a/60610218/4062350