Test some other behaviors subject to bit rot
If not routinely tested, there are some other corners that may be
prone to bit rot, such as:

- make sure DROP EXTENSION works cleanly, and leaves nothing behind.

- make sure LOAD 'pljava-so-...'; works to load as a non-extension
  (that was introduced for use in the pre-extension versions of PG,
  but still covers one possibly important use case: it works when
  you don't have access to install PL/Java in the specific server
  directories that CREATE EXTENSION requires, but you can install in
  some other directory readable by the server, set pljava.module_path
  to correctly locate the jar files, and pass LOAD the correct path
  to the shared object).

- Make sure the after-the-fact CREATE EXTENSION FROM unpackaged; still
  works. PG upstream considers that obsolete and it is removed from
  PG 13 (but there an equivalent two-step sequence, CREATE EXTENSION
  pljava VERSION unpackaged; followed in a new session by ALTER EXTENSION
  pljava UPDATE; will still work). For PL/Java there is still a
  possibility that CREATE EXTENSION fails on the first try if some
  settings aren't right, and shifts to an exploratory installation
  that succeeds when the wrong settings are fixed, and then prints
  the CREATE FROM unpackaged command(s) that will be needed to finish
  the job.

The from-unpackaged extension script had bit-rotted; with the addition
of validators back in a81ef5e, the CREATE LANGUAGE commands needed to
be removed from the script, and the alias_java_language function born
in 5565a3c should have been added.

Update Node.java to record the installed location of the shared object
and offer a loadPLJava() method that issues the right LOAD command.
Other small fixes. Apparently the pgjdbc-ng ResultSetMetaData can
return precision = -1 for some data types, which JDBC does not consider
valid. WebRowSet.populate() takes that in stride and just presents it
as 0 instead, but RowSetMetaDataImpl.setPrecision() rejects it.

On Windows, I think installJar was constructing URIs with Windows
backslashes as separators. And it was working .... With any luck,
constructing them as proper URIs will also work.

These blobs of jshell in the three different CI service configs are
by now getting long enough, and similar enough, to be ripe for
comparing one last time and factoring out one common version, but
that can be a chore for another day.