Add a pljava.policydefault build-time property
This should normally be a string that parses into two elements
separated by a comma, the first a double-quoted file: URL, and
the second a double-quoted equals sign (meaning that PL/Java
should not continue parsing additional policy files from the
base Java security properties). The list can be of length other
than two if circumstances demand; any URL can be prefixed with
n= for some integer n (all inside the double quotes) to start
replacing the base Java policy URLs at that index (the default
is 2, as covered in policy.md), and runtime Java properties can
be expanded in a URL. The default, if this property is not
specified, is

  "file:${org.postgresql.sysconfdir}/pljava.policy","="

Maven needs to *receive* it in that form, so the rules of whatever
shell is invoking Maven will need to be followed to properly quote
that whole thing (with -Dpljava.policydefault= in front). Assuming
Maven receives it intact, how Maven *delivers* it intact from
there to the C compiler will be a respectable test, especially on
Windows, of whether the GSoC 2020 work got the quoting rules right
in the PGXS plugin.