Install 1.6.0's new sqlj function when updating
The addition of a new function in the sqlj schema should have
occasioned a new SchemaVariant in InstallHelper that would make sure
the same function gets installed during an update from an earlier
installed version. That was overlooked.

As the necessary SQL to create the function is exactly the same as
what is generated automatically into the deployment descriptor,
refactor slightly to allow a SchemaVariant's migrateFrom to also
apply deployment-descriptor commands, selectively, by replacing the
default recognized implementor tag set. The new function added
for 1.6.0 is now given a distinct implementor tag so it can be
selectively applied.

A new SQLAction is added to the o.p.p.management.Commands class,
under the default implementor tag, that unconditionally adds the
new tag to the recognized set. So in a full ex-nihilo install,
everything gets installed.

The 1.5 to 1.6 schema migration now simply executes the deployment
descriptor with the default recognized tag set replaced, and containing
only the specific tag for the new function, so only that gets installed.

Addresses #341.