Mollify the -Xcheck:jni exception-check checker
Nearly all JNI calls-into-Java are made through the wrappers in
JNICalls.c, which ensure that each one is followed by the obligatory
was-an-exception-thrown check. But this handful of uses in JNICalls
itself weren't followed by checks. As it happens, the methods being
called wouldn't be expected to throw exceptions. But the -Xcheck:jni
linter doesn't know that, so add the checks here to make it happy.

After 15 years of not having exception checks here, it's certainly
not necessary to implement any elaborate recovery strategies for them,
other than maybe not using the method's result. They're all "shouldn't
happen" cases.