Run clinit at validation in the right context
If validating (check_function_bodies is on), class initializers
will be run at CREATE FUNCTION time in a bid to catch as many
issues early as possible. Run them in the same access control
context that would be used at run time to call the function
being declared.

Other classes besides the one containing the target method can
also be loaded and initialized: all those corresponding to the
method's parameter and return types, for example. Run those class
initializers also under the context that would be used for the
function being declared; that also mirrors what will happen at
runtime when the function is called.

If a class contains several methods that would be given different
access control contexts (will be declared with different trust or
language attributes, say), the permissions available when the class
initializer runs will be those of whichever function is called first
in a given session (or, by the validator, for whichever CREATE
FUNCTION is seen first in a session). If there are initial actions
needed that require permissions that might not always be available,
they should be moved out of the class initializer and handled in
the first call of a method that will have the needed permissions.
Or, perhaps, things can be arranged so that the more-privileged
function will be the first one called in any session, and have its
CREATE FUNCTION first in the deployment descriptor, so that the
class initializer will always run with its permissions.

Addresses #342.

In passing, unwrap ExceptionInInitializerError payloads when
returning from a call; it's more useful to have the ereport
show what actually went wrong.