Give SQLType an optional= element
Because Java disallows null as an annotation element value, it has
never been possible to say defaultValue=null. Allow optional=true
as an alternative meaning just that.

Use it to tweak the PassXML.transformXML function so it doesn't
set indent-related properties on the transformer unless the caller
specified them explicitly.