Fix a bug in preliminary path generation
This code fell behind the addition of synthetic=TWIN (now, until
characterize, synthetic==null in that case, so isSynthetic is the
right way to tell if the operator is synthetic).

With this fixed, the right paths are generated, which can include
multiple ones that differ by commutation and can't be pruned
until characterize time when the operand types have been resolved.

Next needed step is an algorithm to do that characterize-time pruning.

Also, in case of multiple operator names for the same operation,
synthesized functions should be remembered and shared, not duplicated.