Revisit getPgConfigPropertyAsList
The change in f8e475b was a hasty measure on discovering that
the single quotes in a quoted argument reported by pg_config
may appear within the argument rather than only surrounding it.
It solved that problem, but can be fooled into empty arguments
where multiple spaces separate arguments in the pg_config output
(issue #347).

Revert to the regex-based approach from before f8e475b and just
tweak that.

This implementation openly does not handle the case of an argument
containing a single quote as content. PostgreSQL's own build process
failed in a simple test of such a case, so supporting it here is
probably not necessary for now.