Add message if Java VM calls exit() at start
This seems to be the best that can be done if the initial
module layer can't be constructed because one module's declared
dependency can't be resolved. The VM just writes a message to
standard error and immediately exits in that case. Use the
on_proc_exit hook to make sure a logged message results. The VM's
message to standard error only gets to the log if logging_collector
is active.