Avoid segfault from early _destroyJavaVM
This was broken by bbbb75c. Fixing it restores reasonably graceful
reporting when the implementation jar in pljava.module_path isn't
found.

When the API jar is the one that's not found, the reporting is not
so good. "Error occurred during initialization of boot layer" is
reported early and not being caught, followed by an unexplained exit.
The message can be lost if the backend's standard error is not being
collected for the log.