Trigger no sqlj schema groundwork from validator
A thinko in the validator handler sought to avoid one problem
by inviting another; if the validator handler is called before
initialization is complete, its preemptive early setting of
pljavaLoadPath could be interpreted in the initsequencer as
indicating an explicit LOAD or CREATE EXTENSION, calling for
the sqlj schema to be populated or migrated. During a pg_upgrade,
that schema might not be fully populated, foiling the schema
recognition/migration code.

So, rather than preemptively setting the load path, just avoid
the exact problem the code was meant to avoid, in a more direct
manner.

Addresses #352.