Treat an autovacuum worker process specially
PL/Java can be invoked in such a process if there is
a functional index declared that uses a PL/Java function.
Like the background worker case, there may not be a MyProcPort
structure available.

The "can't support background worker before PG 9.5" check was too
tersely explained in 97ddb5e where it was added. While earlier PG
versions maintained an AuthenticatedUserId value, there was no
accessible getter for it before PG 9.5. 9.5 also added the two-argument
form of GetUserNameFromId used here, but if that were the only problem,
it could be handled (the former one-argument behavior was the same as
specified here with second argument false). The inability to obtain
the authenticated user id at all is what's hopeless in earlier versions.

The limitation also applies in the autovacuum worker case. In versions
prior to PG 9.5, it may be necessary to use false as the per-table
autovacuum_enabled storage parameter on any table having a Java-based
functional index, and run explicit VACUUM on such tables over a normal
connection. Prior to PG 8.4, there was no such per-table setting, and
the only workaround would be to forego autovacuum database-wide and
use explicit VACUUM on some schedule.