Detect if running in a post-JEP411 JVM
PL/Java 1.5's build system already breaks as of Java 15, so it
won't be necessary to add SuppressWarnings("removal") anywhere
in the 1.5 sources. It also will not start by default on a JVM
later than 17, because it does not use -Djava.security.manager=allow
among the invocation options by default. The logic here is only to
detect if it has been run with that option explicitly added (in
pljava.vmoptions) and is running on a Java version where the
functionality has been gutted.