Tighten deserialization of Identifier
I don't have any PL/Java application in mind that would involve
serializing identifiers (or Principals containing them), but the
Subject class requires Principals that are serializable, hence
this make-work. (It might be simpler in the long run to implement
another DomainCombiner that doesn't use the Subject class.)

The only serialVersionUID to change will be for Identifier itself.
The other changes merely tighten exclusions of cases that shouldn't
happen anyway.

I'm going to switch to a policy, where a serialVersionUID must change,
of changing to a small integer that will simply increment. The default
calculation makes ugly numbers and offers no advantage if there's
a compatibility break anyway.